# code-review-ai/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CodeReviewAI API
  version: 0.1.0
paths:
...
  /api/v1/knowledgegraph/concepts/{concept_slug}:
    put:
      summary: Upsert Concept
      operationId: upsert_concept_api_v1_knowledgegraph_concepts__concept_slug__put
      parameters:
        - name: concept_slug
          in: path
          required: true
          schema:
            type: string
            title: Concept Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertConceptRequest'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphConceptResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledgegraph/concepts/{concept_slug}/relations:
    patch:
      summary: Patch Concept Relations
      operationId: patch_concept_relations_api_v1_knowledgegraph_concepts__concept_slug__relations_patch
      parameters:
        - name: concept_slug
          in: path
          required: true
          schema:
            type: string
            title: Concept Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchConceptRelationsRequest'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphConceptResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledgegraph/exercises/{exercise_id}:
    put:
      summary: Upsert Exercise
      operationId: upsert_exercise_api_v1_knowledgegraph_exercises__exercise_id__put
      parameters:
        - name: exercise_id
          in: path
          required: true
          schema:
            type: string
            title: Exercise Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertExerciseRequest'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphExerciseResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledgegraph/exercises/{exercise_id}/relations:
    patch:
      summary: Patch Exercise Relations
      operationId: patch_exercise_relations_api_v1_knowledgegraph_exercises__exercise_id__relations_patch
      parameters:
        - name: exercise_id
          in: path
          required: true
          schema:
            type: string
            title: Exercise Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchExerciseRelationsRequest'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphExerciseResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledgegraph/students/{student_id}:
    put:
      summary: Upsert Student
      operationId: upsert_student_api_v1_knowledgegraph_students__student_id__put
      parameters:
        - name: student_id
          in: path
          required: true
          schema:
            type: string
            title: Student Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertStudentProfileRequest'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphStudentResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledgegraph/submissions/{submission_id}:
    put:
      summary: Upsert Submission
      operationId: upsert_submission_api_v1_knowledgegraph_submissions__submission_id__put
      parameters:
        - name: submission_id
          in: path
          required: true
          schema:
            type: string
            title: Submission Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSubmissionRequest'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphSubmissionResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/knowledgegraph/reviews/{review_id}:
    put:
      summary: Upsert Review
      operationId: upsert_review_api_v1_knowledgegraph_reviews__review_id__put
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: string
            title: Review Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertReviewRequest'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeGraphReviewResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ColumnContext:
      properties:
        start:
          anyOf:
            - type: integer
            - type: "null"
          title: Start
        end:
          anyOf:
            - type: integer
            - type: "null"
          title: End
      type: object
      required:
        - start
        - end
      title: ColumnContext
    PatchConceptRelationsRequest:
      properties:
        prerequisite_slugs:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Prerequisite Slugs
      type: object
      title: PatchConceptRelationsRequest
    PatchExerciseRelationsRequest:
      properties:
        concept_slugs:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Concept Slugs
        related_exercise_ids:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Related Exercise Ids
        related_exercise_slugs:
          anyOf:
            - items:
                type: string
              type: array
            - type: "null"
          title: Related Exercise Slugs
      type: object
      title: PatchExerciseRelationsRequest
    ConceptRecord:
      properties:
        concept_id:
          type: string
          title: Concept Id
        slug:
          type: string
          title: Slug
          default: ""
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ""
        difficulty:
          type: integer
          title: Difficulty
          default: 1
      type: object
      required:
        - concept_id
        - name
      title: ConceptRecord
...
    ExerciseRecord:
      properties:
        exercise_id:
          type: string
          title: Exercise Id
        slug:
          type: string
          title: Slug
          default: ""
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        content:
          type: string
          title: Content
        difficulty:
          type: string
          title: Difficulty
        tags:
          items:
            type: string
          type: array
          title: Tags
      type: object
      required:
        - exercise_id
        - title
        - description
        - content
        - difficulty
      title: ExerciseRecord
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    KnowledgeGraphConceptResponse:
      properties:
        concept:
          $ref: '#/components/schemas/ConceptRecord'
      type: object
      required:
        - concept
      title: KnowledgeGraphConceptResponse
...
    KnowledgeGraphExerciseResponse:
      properties:
        exercise:
          $ref: '#/components/schemas/ExerciseRecord'
      type: object
      required:
        - exercise
      title: KnowledgeGraphExerciseResponse
    KnowledgeGraphReviewResponse:
      properties:
        review:
          $ref: '#/components/schemas/ReviewRecord'
      type: object
      required:
        - review
      title: KnowledgeGraphReviewResponse
...
    KnowledgeGraphStudentResponse:
      properties:
        student_id:
          type: string
          title: Student Id
        student_profile:
          $ref: '#/components/schemas/StudentProfileScoring'
      type: object
      required:
        - student_id
        - student_profile
      title: KnowledgeGraphStudentResponse
    KnowledgeGraphSubmissionResponse:
      properties:
        submission:
          $ref: '#/components/schemas/SubmissionRecord'
      type: object
      required:
        - submission
      title: KnowledgeGraphSubmissionResponse
    LineContext:
      properties:
        start:
          anyOf:
            - type: integer
            - type: "null"
          title: Start
        end:
          anyOf:
            - type: integer
            - type: "null"
          title: End
      type: object
      required:
        - start
        - end
      title: LineContext
...
    ReviewItem:
      properties:
        line:
          $ref: '#/components/schemas/LineContext'
        column:
          anyOf:
            - $ref: '#/components/schemas/ColumnContext'
            - type: "null"
        code_snippet:
          type: string
          title: Code Snippet
        type:
          type: string
          enum:
            - Error
            - Warning
          title: Type
        issue:
          type: string
          title: Issue
          description: For a 'Warning', this should explain a clean-code, readability, or refactoring improvement in simple terms.
        fix_suggestion:
          type: string
          title: Fix Suggestion
        review_link:
          anyOf:
            - $ref: '#/components/schemas/ReviewLink'
            - type: "null"
      type: object
      required:
        - code_snippet
        - type
        - issue
        - fix_suggestion
      title: ReviewItem
    ReviewLink:
      properties:
        previous_submission_id:
          type: string
          title: Previous Submission Id
        previous_code_snippets:
          items:
            type: string
          type: array
          title: Previous Code Snippets
        comparison_mode:
          anyOf:
            - type: string
            - type: "null"
          title: Comparison Mode
        what_improved:
          type: string
          title: What Improved
        what_still_needs_work:
          type: string
          title: What Still Needs Work
        relation_summary:
          type: string
          title: Relation Summary
      type: object
      required:
        - previous_submission_id
        - what_improved
        - what_still_needs_work
        - relation_summary
      title: ReviewLink
    ReviewRecord:
      properties:
        review_id:
          type: string
          title: Review Id
        student_id:
          type: string
          title: Student Id
        exercise_id:
          type: string
          title: Exercise Id
          default: ""
        submission_id:
          type: string
          title: Submission Id
          default: ""
        current_concept:
          type: string
          title: Current Concept
          default: ""
        created_at:
          type: string
          title: Created At
          default: ""
        summary:
          type: string
          title: Summary
        detail:
          type: string
          title: Detail
      type: object
      required:
        - review_id
        - student_id
        - summary
        - detail
      title: ReviewRecord
...
    ScoreCard:
      properties:
        problem_solving_creativity:
          $ref: '#/components/schemas/ScoreCardItem'
        logic_traceability:
          $ref: '#/components/schemas/ScoreCardItem'
        generalization_score:
          $ref: '#/components/schemas/ScoreCardItem'
        construct_appropriateness:
          $ref: '#/components/schemas/ScoreCardItem'
        self_correction_path:
          $ref: '#/components/schemas/ScoreCardItem'
        variable_understanding:
          $ref: '#/components/schemas/ScoreCardItem'
        control_flow_understanding:
          $ref: '#/components/schemas/ScoreCardItem'
        input_output_awareness:
          $ref: '#/components/schemas/ScoreCardItem'
        edge_case_awareness:
          $ref: '#/components/schemas/ScoreCardItem'
        debugging_readiness:
          $ref: '#/components/schemas/ScoreCardItem'
      type: object
      required:
        - problem_solving_creativity
        - logic_traceability
        - generalization_score
        - construct_appropriateness
        - self_correction_path
        - variable_understanding
        - control_flow_understanding
        - input_output_awareness
        - edge_case_awareness
        - debugging_readiness
      title: ScoreCard
...
    StudentProfileScoring:
      properties:
        concept_mastery:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Concept Mastery
        implementation_consistency:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Implementation Consistency
        debugging_independence:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Debugging Independence
        efficiency_awareness:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Efficiency Awareness
        concept_transfer:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Concept Transfer
        learning_velocity:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Learning Velocity
        notes:
          type: string
          title: Notes
          default: ""
      type: object
      required:
        - concept_mastery
        - implementation_consistency
        - debugging_independence
        - efficiency_awareness
        - concept_transfer
        - learning_velocity
      title: StudentProfileScoring
...
    SubmissionRecord:
      properties:
        submission_id:
          type: string
          title: Submission Id
        student_id:
          type: string
          title: Student Id
        exercise_id:
          type: string
          title: Exercise Id
          default: ""
        code:
          type: string
          title: Code
        testcase_outputs:
          items:
            $ref: '#/components/schemas/SubmissionTestCaseOutput'
          type: array
          title: Testcase Outputs
        created_at:
          type: string
          title: Created At
          default: ""
      type: object
      required:
        - submission_id
        - student_id
        - code
      title: SubmissionRecord
...
    SubmissionTestCaseOutput:
      properties:
        expect:
          type: string
          title: Expect
          default: ""
        output:
          type: string
          title: Output
          default: ""
      type: object
      title: SubmissionTestCaseOutput
...
    UpsertConceptRequest:
      properties:
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
          default: ""
        difficulty:
          type: integer
          title: Difficulty
          default: 1
      type: object
      required:
        - name
      title: UpsertConceptRequest
    UpsertExerciseRequest:
      properties:
        slug:
          type: string
          title: Slug
          default: ""
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        content:
          type: string
          title: Content
        difficulty:
          type: string
          title: Difficulty
        tags:
          items:
            type: string
          type: array
          title: Tags
      type: object
      required:
        - title
        - description
        - content
        - difficulty
      title: UpsertExerciseRequest
    UpsertReviewRequest:
      properties:
        submission_id:
          type: string
          title: Submission Id
        summary:
          type: string
          title: Summary
        detail:
          type: string
          title: Detail
        review_items:
          items:
            $ref: '#/components/schemas/ReviewItem'
          type: array
          title: Review Items
        scorecard:
          $ref: '#/components/schemas/ScoreCard'
        current_concept:
          type: string
          title: Current Concept
          default: ""
      type: object
      required:
        - submission_id
        - summary
        - detail
        - scorecard
      title: UpsertReviewRequest
    UpsertStudentProfileRequest:
      properties:
        student_profile:
          $ref: '#/components/schemas/StudentProfileScoring'
      type: object
      required:
        - student_profile
      title: UpsertStudentProfileRequest
    UpsertSubmissionRequest:
      properties:
        student_id:
          type: string
          title: Student Id
        exercise_id:
          type: string
          title: Exercise Id
        code:
          type: string
          title: Code
        testcase_outputs:
          items:
            $ref: '#/components/schemas/SubmissionTestCaseOutput'
          type: array
          title: Testcase Outputs
      type: object
      required:
        - student_id
        - exercise_id
        - code
      title: UpsertSubmissionRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError